pyproject.toml
[tool.uv.workspace]
members = [".", "examples/fastapi"]

[project]
name = "py-identity-model"
version = "1.2.0"
description = "OAuth2.0 and OpenID Connect Client Library"
authors = [{ name = "jamescrowley321", email = "jamescrowley151@gmail.com" }]
requires-python = ">=3.12"
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "PyJWT>=2.9.0,<3",
    "httpx>=0.28.1,<1",
    "cryptography>=45.0.2,<46",
    "async-lru>=2.0.4,<3",
]

[dependency-groups]
dev = [
    "pre-commit>=3.8.0,<4",
    "python-dotenv>=1.0.1,<2",
    "pytest>=8.3.2,<9",
    "pytest-asyncio>=0.24.0",
    "respx>=0.21.0",
    "pytest-timeout>=2.3.1",
    "ruff>=0.11.12",
    "pyrefly>=0.34.0"
]
docs = [
    "mkdocs>=1.6.0,<2",
    "mkdocs-material>=9.6.0,<10",
    "mkdocs-awesome-pages-plugin>=2.10.0,<3"
]

[build-system]
requires = ["uv_build>=0.6,<0.7"]
build-backend = "uv_build"

[tool.semantic_release]
assets = []
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "angular"
logging_use_named_masks = false
major_on_zero = true
tag_format = "{version}"
build_command = "echo 1"
version_toml = [
    "pyproject.toml:project.version",
]
commit_version_number = true

[tool.semantic_release.branches.main]
match = "main"

[tool.semantic_release.branches.other]
match = ".*"
prerelease_token = "rc"
prerelease = true

[tool.semantic_release.changelog]
changelog_file = "CHANGELOG.md"
exclude_commit_patterns = []

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = true

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["build", "chore", "ci", "docs", "feat", "fix", "perf", "style", "refactor", "test"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true

[tool.ruff]
line-length = 79
target-version = "py312"

[tool.ruff.lint]
select = [
    "B", # flake8-bugbear
    "C", # mccabe complexity
    "E", # pycodestyle errors
    "F", # pyflakes
    "W", # pycodestyle warnings
    "I", # isort
    "UP", # pyupgrade - modernize Python code
    "S", # flake8-bandit - security checks
    "A", # flake8-builtins - check for shadowing builtins
    "COM", # flake8-commas - trailing commas
    "DTZ", # flake8-datetimez - datetime timezone awareness
    "ICN", # flake8-import-conventions - import naming conventions
    "PIE", # flake8-pie - misc lints
    "PT", # flake8-pytest-style - pytest best practices
    "Q", # flake8-quotes - quote consistency
    "RSE", # flake8-raise - exception raise improvements
    "RET", # flake8-return - return statement improvements
    "SIM", # flake8-simplify - code simplification suggestions
    "TCH", # flake8-type-checking - optimize type checking imports
    "ARG", # flake8-unused-arguments - unused function arguments
    "PTH", # flake8-use-pathlib - prefer pathlib over os.path
    "ERA", # eradicate - find commented out code
    "PL", # pylint
    "PERF", # perflint - performance anti-patterns
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",      # line too long (handled by formatter)
    "COM812",    # trailing comma missing (conflicts with formatter)
    "ISC001",    # implicit string concatenation (conflicts with formatter)
    "S101",      # use of assert (needed for tests)
    "S104",      # possible binding to all interfaces
    "S105",      # possible hardcoded password
    "S106",      # possible hardcoded password
    "S113",      # probable use of requests call without timeout
    "B008",      # function calls in argument defaults (FastAPI pattern)
    "PLR0911",   # too many return statements
    "PLR0913",   # too many arguments
    "PLR2004",   # magic value used in comparison
    "PLC0415",   # import should be at top-level (some are intentional)
    "ARG001",    # unused function argument (e.g., FastAPI Request)
    "RUF003",    # ambiguous unicode character in comment
    "PERF401",   # use list comprehension (less readable in some cases)
    "PERF403",   # use dict comprehension (less readable in some cases)
    "RET504",    # unnecessary assignment before return (improves readability)
    "A001",      # variable shadowing builtin (format is acceptable)
    "A002",      # argument shadowing builtin (format is acceptable)
    "PIE796",    # enum duplicate values (intentional for constants)
    "PT011",     # pytest.raises too broad (acceptable for integration tests)
    "DTZ005",    # datetime without timezone (acceptable for benchmarks)
    "PLW0603",   # global statement (acceptable for test fixtures)
    "PTH113",    # use pathlib (os.path is acceptable)
    "PLR1722",   # use sys.exit (exit is acceptable)
    "RUF043",    # regex metacharacters (not actual regex patterns)
]

[tool.ruff.lint.isort]
known-first-party = ["py_identity_model"]
force-sort-within-sections = true
lines-after-imports = 2

[tool.ruff.lint.mccabe]
max-complexity = 18

[tool.ruff.lint.pyupgrade]
# Keep runtime type checking for Python 3.12+
keep-runtime-typing = true

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Indent with spaces
indent-style = "space"
# Respect magic trailing comma
skip-magic-trailing-comma = false
# Auto-detect line ending
line-ending = "auto"
# Format docstrings
docstring-code-format = true

[tool.pytest.ini_options]
# Test discovery
testpaths = ["src/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# Output options
addopts = [
    "-ra", # show extra test summary info for all except passed
    "--strict-markers", # raise error on unknown markers
    "--strict-config", # raise error on config issues
    "--showlocals", # show local variables in tracebacks
    "-vv", # very verbose
]
# Markers for test categorization
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "unit: marks tests as unit tests",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
# Async test configuration
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
# Filter warnings
filterwarnings = [
    "error", # treat warnings as errors
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src/py_identity_model"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
